+++ src/core.js
@@ -1,4 +1,5 @@
 import propertyValueConverters from './internal/property-value-converters'
 
+export * from './internal/convert'
 export * from './internal/utils'
 export {propertyValueConverters}
